fix passing themes into blocks
